requirements.txt
fastapi
jose
pandas
prometheus_fastapi_instrumentator
protobuf
pydantic[email]
pydantic_settings
python_jose
starlette
uvicorn
google-auth
requests
sqlalchemy
psycopg2
qrcode
Pillow
httpx